.teamcity/_self/projects/WebApp.kt - move CalypsoPreReleaseDashboard build config into WebApp.kt
.teamcity/settings.kts
- change "CALYPSO_E2E_DASHBOARD_AWS_S3_ROOT" type to text

.teamcity/_self/projects/CalypsoDashboard.kt
- remvoe
